+++ teamtalk/permission.py
@@ -8,6 +8,9 @@ class _PermissionMeta(type):
         name = f"USERRIGHT_{name}"
         return getattr(sdk.UserRight, name, None)
 
+    def __dir__(cls) -> list[str]:
+        return [name[10:] for name in dir(sdk.UserRight) if name.startswith("USERRIGHT_")]
+
 
 class Permission(metaclass=_PermissionMeta):
     """A class representing user permissions in TeamTalk."""