+++ teamtalk/subscription.py
@@ -8,6 +8,9 @@ class _SubscriptionMeta(type):
         name = f"SUBSCRIBE_{name}"
         return getattr(sdk.Subscription, name, None)
 
+    def __dir__(cls) -> list[str]:
+        return [name[10:] for name in dir(sdk.Subscription) if name.startswith("SUBSCRIBE_")]
+
 
 class Subscription(metaclass=_SubscriptionMeta):
     """A class representing subscriptions in TeamTalk."""